[tests] Consolidate more imports
