+++ tests/e2e/api/coin.test.ts
@@ -1,7 +1,6 @@
-import { AptosConfig, Network, Aptos, Account, Deserializer } from "../../../src";
+import { AptosConfig, Network, Aptos, Account, Deserializer, TypeTagStruct } from "../../../src";
 import { waitForTransaction } from "../../../src/internal/transaction";
 import { RawTransaction, TransactionPayloadEntryFunction } from "../../../src/transactions/instances";
-import { TypeTagStruct } from "../../../src/transactions/typeTag/typeTag";
 import { sleep } from "../../../src/utils/helpers";
 import { FUND_AMOUNT, INDEXER_WAIT_TIME } from "../../unit/helper";
