+++ tests/unit/script_transaction_arguments.test.ts
@@ -1,9 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Serializer, Deserializer } from "../../src/bcs";
-import { AccountAddress } from "../../src/core";
-import { Bool, U128, U16, U256, U32, U64, U8 } from "../../src/bcs/serializable/move-primitives";
+import { AccountAddress, Serializer, Deserializer, Bool, U128, U16, U256, U32, U64, U8 } from "../../src";
 import { MoveVector } from "../../src/bcs/serializable/move-structs";
 import { ScriptFunctionArgument, deserializeFromScriptArgument } from "../../src/transactions/instances";
